Нода вынесена в отдельный файл
